refactor: separate chat context menu files in folder
